Do not start in recents tab if we do not have recent emojis
If user has not used emojis we should not bring recents emoji tab on the
first invocation of the on-screen keyboard as it is confusing to see an
empty tab.

We will still be showing recents tab even if it is empty if user
actively switched to it when using on-screen keyboard.

Bug: 24497650

Change-Id: I68b4bf64441590858d246168cc4733a2875a1357
(cherry picked from commit 24d67eba27cb23c0d3aca44236247ec12380daa5)